Bump 1.0.0-beta8: 12 项修复 + 国内构建网络优化
修复（双端同步）：
- 燃料重复注入漏洞（高）：Simulate 先确认、Execute 不再二次注入
- 能量条渲染 int 溢出（中）：改用 long 计算渲染比例
- 菜单数据 int 截断溢出：4 个数据源改用大容量类型传输
- GUI 菜单客户端不可用：stillValid() 双端一致性修正
- 涡轮机 getRotorComponentTypeAt() Y 轴墙体误判
- getReactorHeat() NoSuchElementException：空列表 getFirst() 崩溃
- 废物条使用燃料容量代替废物容量
- 线圈 tag 解析错误：线圈数量恒为 0
- onBlockRemoved() 未清理 _pendingControllerTag
- adjustControlRod 快速点击竞态：模拟阶段先行确认
- 控制棒英文文本超出 GUI 边界
- 反应堆屏幕添加废物量标签

构建优化：
- build.gradle：增加 Aliyun Maven Central 镜像（20MB/s）
- 移除 maven.neoforged.net 官方仓库（国内 SSL 连接被重置）
- afterEvaluate 钩子移除 NeoGradle 插件自动追加的失效仓库
- .gitignore 添加 server-*.log 排除 dev 日志

+++ versions/1.21.1/src/main/java/com/compact/extremereactor/client/screen/CompactTurbineScreen.java
@@ -46,7 +46,7 @@ public class CompactTurbineScreen extends AbstractContainerScreen<CompactTurbine
     private void renderVerticalBar(GuiGraphics guiGraphics, int x, int y, int width, int height, int value, int capacity, int color) {
         guiGraphics.fill(x, y, x + width, y + height, 0xFF000000);
         if (capacity > 0 && value > 0) {
-            final int filled = Math.min(height - 2, (height - 2) * value / capacity);
+            final int filled = Math.min(height - 2, (int)((long)(height - 2) * value / capacity));
             guiGraphics.fill(x + 1, y + height - 1 - filled, x + width - 1, y + height - 1, color);
         }
     }
@@ -55,7 +55,7 @@ public class CompactTurbineScreen extends AbstractContainerScreen<CompactTurbine
     private void renderHorizontalBar(GuiGraphics guiGraphics, int x, int y, int width, int height, int value, int capacity, int color) {
         guiGraphics.fill(x, y, x + width, y + height, 0xFF000000);
         if (capacity > 0 && value > 0) {
-            final int filled = Math.min(width - 2, (width - 2) * value / capacity);
+            final int filled = Math.min(width - 2, (int)((long)(width - 2) * value / capacity));
             guiGraphics.fill(x + 1, y + 1, x + 1 + filled, y + height - 1, color);
         }
     }
